+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 
 ### Fixed
 
+- Public: Fix "File type not supported" error on snapshot upload in selfie step.
+
 ## [6.4.0]
 
 ### Added